+++ sentry/src/main/java/io/sentry/SentryReplayOptions.java
@@ -11,6 +11,12 @@ public final class SentryReplayOptions {
 
   public static final String TEXT_VIEW_CLASS_NAME = "android.widget.TextView";
   public static final String IMAGE_VIEW_CLASS_NAME = "android.widget.ImageView";
+  public static final String WEB_VIEW_CLASS_NAME = "android.webkit.WebView";
+  public static final String VIDEO_VIEW_CLASS_NAME = "android.widget.VideoView";
+  public static final String ANDROIDX_MEDIA_VIEW_CLASS_NAME = "androidx.media3.ui.PlayerView";
+  public static final String EXOPLAYER_CLASS_NAME = "com.google.android.exoplayer2.ui.PlayerView";
+  public static final String EXOPLAYER_STYLED_CLASS_NAME =
+      "com.google.android.exoplayer2.ui.StyledPlayerView";
 
   public enum SentryReplayQuality {
     /** Video Scale: 80% Bit Rate: 50.000 */
@@ -99,6 +105,11 @@ public final class SentryReplayOptions {
   public SentryReplayOptions() {
     setMaskAllText(true);
     setMaskAllImages(true);
+    maskViewClasses.add(WEB_VIEW_CLASS_NAME);
+    maskViewClasses.add(VIDEO_VIEW_CLASS_NAME);
+    maskViewClasses.add(ANDROIDX_MEDIA_VIEW_CLASS_NAME);
+    maskViewClasses.add(EXOPLAYER_CLASS_NAME);
+    maskViewClasses.add(EXOPLAYER_STYLED_CLASS_NAME);
   }
 
   public SentryReplayOptions(